feat: 전체 뉴스 창 추가
