+++ Targets/D3N/Sources/Feature/AllNews/AllNewsNavigationStackStore.swift
@@ -1,16 +1,16 @@
 //
-//  TodayNavigationStackStore.swift
+//  AllNewsNavigationStackStore.swift
 //  D3N
 //
-//  Created by 송영모 on 10/12/23.
+//  Created by Younghoon Ahn on 11/14/23.
 //  Copyright © 2023 sju. All rights reserved.
 //
 
 import ComposableArchitecture
 
-public struct TodayNavigationStackStore: Reducer {
+public struct AllNewsNavigationStackStore: Reducer {
     public struct State: Equatable {
-        var main: TodayMainStore.State = .init()
+        var main: AllNewsStore.State = .init()
         
         var path: StackState<Path.State> = .init()
     }
@@ -20,7 +20,7 @@ public struct TodayNavigationStackStore: Reducer {
         
         case onAppear
         
-        case main(TodayMainStore.Action)
+        case main(AllNewsStore.Action)
         case path(StackAction<Path.State, Path.Action>)
         
         case popToRoot
@@ -33,23 +33,18 @@ public struct TodayNavigationStackStore: Reducer {
     
     public struct Path: Reducer {
         public enum State: Equatable {
-            case detail(TodayDetailStore.State)
             case quizMain(QuizMainStore.State)
             case quizResult(QuizResultStore.State)
             case newsList(AllNewsStore.State)
         }
         
         public enum Action: Equatable {
-            case detail(TodayDetailStore.Action)
             case quizMain(QuizMainStore.Action)
             case quizResult(QuizResultStore.Action)
             case newsList(AllNewsStore.Action)
         }
         
         public var body: some Reducer<State, Action> {
-            Scope(state: /State.detail, action: /Action.detail) {
-                TodayDetailStore()
-            }
             Scope(state: /State.quizMain, action: /Action.quizMain) {
                 QuizMainStore()
             }
@@ -74,9 +69,6 @@ public struct TodayNavigationStackStore: Reducer {
                 switch action {
                 case let .select(newsEntity):
                     state.path.append(.quizMain(.init(newsEntity: newsEntity)))
-                    return .none
-                case .allNewsButtonTapped:
-                    state.path.append(.newsList(.init()))
                     return .none
                 }
                 
@@ -109,10 +101,11 @@ public struct TodayNavigationStackStore: Reducer {
             }
         }
         Scope(state: \.main, action: /Action.main) {
-            TodayMainStore()
+            AllNewsStore()
         }
         .forEach(\.path, action: /Action.path) {
             Path()
         }
     }
 }
+
