+++ Targets/D3N/Sources/Feature/AllNews/AllNewsNavigationStackView.swift
@@ -1,8 +1,8 @@
 //
-//  TodayNavigationStackView.swift
+//  AllNewsNavigationStackView.swift
 //  D3N
 //
-//  Created by 송영모 on 10/12/23.
+//  Created by Younghoon Ahn on 11/14/23.
 //  Copyright © 2023 sju. All rights reserved.
 //
 
@@ -11,49 +11,42 @@ import SwiftUI
 
 import ComposableArchitecture
 
-public struct TodayNavigationStackView: View {
-    let store: StoreOf<TodayNavigationStackStore>
+public struct AllNewsNavigationStackView: View {
+    let store: StoreOf<AllNewsNavigationStackStore>
     
-    public init(store: StoreOf<TodayNavigationStackStore>) {
+    public init(store: StoreOf<AllNewsNavigationStackStore>) {
         self.store = store
     }
     
     public var body: some View {
         NavigationStackStore(self.store.scope(
             state: \.path,
-            action: TodayNavigationStackStore.Action.path)) {
+            action: AllNewsNavigationStackStore.Action.path)) {
                 WithViewStore(self.store, observe: { $0 }) { viewStore in
-                    TodayMainView(store: self.store.scope(state: \.main, action: TodayNavigationStackStore.Action.main))
+                    AllNewsView(store: self.store.scope(state: \.main, action: AllNewsNavigationStackStore.Action.main))
                         .onAppear {
                             viewStore.send(.onAppear)
                         }
                 }
             } destination: {
                 switch $0 {
-                case .detail:
-                    CaseLet(
-                        /TodayNavigationStackStore.Path.State.detail,
-                         action: TodayNavigationStackStore.Path.Action.detail,
-                         then: TodayDetailView.init(store:)
-                    )
-                    
                 case .quizMain:
                     CaseLet(
-                        /TodayNavigationStackStore.Path.State.quizMain,
-                         action: TodayNavigationStackStore.Path.Action.quizMain,
+                        /AllNewsNavigationStackStore.Path.State.quizMain,
+                         action: AllNewsNavigationStackStore.Path.Action.quizMain,
                          then: QuizMainView.init(store:)
                     )
                     
                 case .quizResult:
                     CaseLet(
-                        /TodayNavigationStackStore.Path.State.quizResult,
-                         action: TodayNavigationStackStore.Path.Action.quizResult,
+                        /AllNewsNavigationStackStore.Path.State.quizResult,
+                         action: AllNewsNavigationStackStore.Path.Action.quizResult,
                          then: QuizResultView.init(store:)
                     )
                     
                 case .newsList:
                     CaseLet(
-                        /TodayNavigationStackStore.Path.State.newsList,
+                        /AllNewsNavigationStackStore.Path.State.newsList,
                          action: TodayNavigationStackStore.Path.Action.newsList,
                          then: AllNewsView.init(store:)
                     )
